backend/pyproject.toml
[project]
name = "bhl-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Viktor Kulyk",email = "kulikvitek2005@gmail.com"}
]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "langchain (>=1.1.0,<2.0.0)",
    "langchain-anthropic (>=1.2.0,<2.0.0)",
    "langchain-openai (>=1.1.0,<2.0.0)",
    "langchain-groq (>=1.1.0,<2.0.0)",
    "langchain-core (>=1.1.0,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "fastapi (>=0.122.0,<0.123.0)",
    "uvicorn[standard] (>=0.38.0,<0.39.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pydantic (>=2.12.5,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
